requirements.txt
flask
kokoro-onnx
onnxruntime
soundfile
numpy
fitz
pymupdf
python-docx
werkzeug
